Moved observation space to abs
